+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
 * Add `epstein_word_count` script
 * Add `--output-json-files` option
+* Deploy the collected JSON files at https://michelcrypt4d4mus.github.io/epstein_text_messages/json_files_from_epstein_files_nov_2025.json
 
 ### 1.0.5
 * Remove --pickled argument